% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter.R
\name{predict_cl}
\alias{predict_cl}
\title{A Predict_Cluster Function}
\usage{
predict_cl(input)
}
\arguments{
\item{input}{input data in csv or dataframe format}
}
\description{
This function allows you to predict the cluster group in which point lies
}
\examples{
predict_cl()
}
\keyword{file}
\keyword{input}
